pyproject.toml
[project]
name = "rag-milvus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "ollama>=0.6.1",
    "openai>=2.15.0",
    "pymilvus[milvus-lite]>=2.6.6",
    "python-dotenv>=1.2.1",
]
